combat-logs-2
bit naughty, but I've gone and done a fix to the scout AI which was just messy (don't know if it was ACTUALLY causing bugs, but looked bad.

Also added in a delay at the end of turns. And also removed the dependency on 'turn' from AI, and set it to check if the unit is initiative.currentunit. I should do this throughout.

Finally stopped the action UI from freaking out when there was no current unit.